backend/requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
FastAPI
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
networkx
nptyping==2.5.0
numpy
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
uvicorn
tqdm
pymupdf
jsonpickle
Pillow
python-multipart